refactor: split onboarding and api helpers
